gem commands now lazy loaded
moved commands from lib/rubygems/commands/<name> to lib/rubygems/commands/<name>_command, so rubygems will
be able to auto require them when the command is run, rather than on "require 'rubygems'"

changed lib/rubygems_plugin.rb so it only registers the commands, and does not require them as rubygems
command manager now does it.

removed lib/gemcutter.rb since it no longer does anything.

changed tests to pull the command files directly so it is not necessary to install the gem to test it.